+++ MapboxNavigation/String.swift
@@ -1,8 +1,16 @@
 import Foundation
 
+let ISO8601Formatter: DateFormatter = {
+    let formatter = DateFormatter()
+    formatter.locale = Locale(identifier: "en_US_POSIX")
+    formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+    formatter.timeZone = TimeZone(secondsFromGMT: 0)
+    return formatter
+}()
+
 extension String {
     var ISO8601Date: Date? {
-        return Date.ISO8601Formatter.date(from: self)
+        return ISO8601Formatter.date(from: self)
     }
     
     /**
